LightOptionParser: multiple options with arguments.
Previously if there was an option which took multiple arguments, the parser
always returned a flat array of all values given for that option, regardless of
how many arguments it expected. E.g. the following:

    cmd --file-pair f1 f2 --file-pair a b

resulted in the following:

    file-pair: [ f1, f2, a, b ]

For the case where the option is defined as taking only one argument, this
behavior is unchanged, but for the cases where the option is defined as taking
more than one option, an array of arrays is returned instead. For the above
example the result is now:

    file-pair: [ [f1, f2], [a, b] ]

The behavior is the same for options with variable arguments. The following:

    cmd --set f1 f2 --set a b c --set last

results in:

    file-pair: [ [f1, f2], [a, b, c], [last] ]